+++ lib/model/additional_data/additional_data.dart
@@ -11,11 +11,11 @@ import 'package:autonomy_flutter/model/additional_data/view_new_message.dart';
 import 'package:autonomy_flutter/model/additional_data/view_postcard.dart';
 import 'package:autonomy_flutter/util/john_gerrard_helper.dart';
 import 'package:autonomy_flutter/util/log.dart';
-import 'package:autonomy_flutter/util/notification_type.dart';
+import 'package:autonomy_flutter/util/notifications/notification_type.dart';
 import 'package:flutter/cupertino.dart';
 
 class AdditionalData {
-  final NotificationType notificationType;
+  final NotificationDataType notificationType;
   final String? announcementContentId;
   final String? linkText;
 
@@ -31,7 +31,7 @@ class AdditionalData {
     final notificationContentId = json['notification_content_id'];
     try {
       final notificationType =
-          NotificationType.fromString(type ?? json['notification_type']);
+          NotificationDataType.fromString(type ?? json['notification_type']);
       final String? linkText = json['link_text'];
 
       final defaultAdditionalData = AdditionalData(
@@ -41,8 +41,8 @@ class AdditionalData {
       );
 
       switch (notificationType) {
-        case NotificationType.customerSupportNewMessage:
-        case NotificationType.customerSupportCloseIssue:
+        case NotificationDataType.customerSupportNewMessage:
+        case NotificationDataType.customerSupportCloseIssue:
           final issueId = json['issue_id'];
           if (issueId == null) {
             log.warning('AdditionalData: issueId is null');
@@ -54,15 +54,15 @@ class AdditionalData {
             announcementContentId: notificationContentId,
             linkText: linkText,
           );
-        case NotificationType.artworkCreated:
-        case NotificationType.artworkReceived:
-        case NotificationType.galleryNewNft:
+        case NotificationDataType.artworkCreated:
+        case NotificationDataType.artworkReceived:
+        case NotificationDataType.galleryNewNft:
           return view_collection_handler.ViewCollection(
             notificationType: notificationType,
             announcementContentId: notificationContentId,
             linkText: linkText,
           );
-        case NotificationType.newMessage:
+        case NotificationDataType.newMessage:
           final groupId = json['group_id'];
           if (groupId == null) {
             log.warning('AdditionalData: groupId is null');
@@ -74,8 +74,8 @@ class AdditionalData {
             announcementContentId: notificationContentId,
             linkText: linkText,
           );
-        case NotificationType.newPostcardTrip:
-        case NotificationType.postcardShareExpired:
+        case NotificationDataType.newPostcardTrip:
+        case NotificationDataType.postcardShareExpired:
           final indexID = json['indexID'];
           if (indexID == null) {
             log.warning('AdditionalData: indexID is null');
@@ -87,7 +87,7 @@ class AdditionalData {
             announcementContentId: notificationContentId,
             linkText: linkText,
           );
-        case NotificationType.jgCrystallineWorkHasArrived:
+        case NotificationDataType.jgCrystallineWorkHasArrived:
           final jgExhibitionId = JohnGerrardHelper.exhibitionID;
           return ViewExhibitionData(
             exhibitionId: jgExhibitionId ?? '',
@@ -95,7 +95,7 @@ class AdditionalData {
             announcementContentId: notificationContentId,
             linkText: linkText,
           );
-        case NotificationType.jgCrystallineWorkGenerated:
+        case NotificationDataType.jgCrystallineWorkGenerated:
           final tokenId = json['token_id'];
           if (tokenId == null) {
             log.warning('AdditionalData: tokenId is null');
@@ -107,9 +107,9 @@ class AdditionalData {
             announcementContentId: notificationContentId,
             linkText: linkText,
           );
-        case NotificationType.exhibitionViewingOpening:
-        case NotificationType.exhibitionSalesOpening:
-        case NotificationType.exhibitionSaleClosing:
+        case NotificationDataType.exhibitionViewingOpening:
+        case NotificationDataType.exhibitionSalesOpening:
+        case NotificationDataType.exhibitionSaleClosing:
           final exhibitionId = json['exhibition_id'];
           if (exhibitionId == null) {
             log.warning('AdditionalData: exhibitionId is null');
@@ -121,7 +121,7 @@ class AdditionalData {
             announcementContentId: notificationContentId,
             linkText: linkText,
           );
-        case NotificationType.navigate:
+        case NotificationDataType.navigate:
           final navigationRoute = json['navigation_route'];
           final homeIndex = json['home_index'];
           return NavigateAdditionalData(
@@ -131,7 +131,7 @@ class AdditionalData {
             homeIndex: homeIndex,
             linkText: linkText,
           );
-        case NotificationType.dailyArtworkReminders:
+        case NotificationDataType.dailyArtworkReminders:
           final subType = json['notification_sub_type'];
           final dailyType = DailyNotificationType.fromString(subType ?? '');
           if (dailyType == null) {
@@ -151,7 +151,7 @@ class AdditionalData {
     } catch (_) {
       log.info('AdditionalData: error parsing additional data');
       return AdditionalData(
-          notificationType: NotificationType.general,
+          notificationType: NotificationDataType.general,
           announcementContentId: notificationContentId);
     }
   }
