+++ lib/screen/settings/preferences/preferences_bloc.dart
@@ -16,7 +16,7 @@ import 'package:autonomy_flutter/service/local_auth_service.dart';
 import 'package:autonomy_flutter/service/settings_data_service.dart';
 import 'package:autonomy_flutter/util/biometrics_util.dart';
 import 'package:autonomy_flutter/util/log.dart';
-import 'package:autonomy_flutter/util/notification_util.dart';
+import 'package:autonomy_flutter/util/notifications/notification_util.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:local_auth/local_auth.dart';
 import 'package:permission_handler/permission_handler.dart';
